move css file under directory
