+++ package.json
@@ -34,7 +34,6 @@
   ],
   "devDependencies": {
     "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.2.2",
-    "webpack-dev-server": "^3.11.0"
+    "@babel/core": "^7.2.2"
   }
 }
