+++ src/component/d3Charts/responsive/response.js
@@ -13,7 +13,6 @@ export default ChartComponent => (
     }
 
     componentDidMount() {
-      console.log('mounted')
       this.fitParentContainer()
       window.addEventListener('resize', this.fitParentContainer)
     }
@@ -26,7 +25,8 @@ export default ChartComponent => (
       const { containerWidth } = this.state
       const currentContainerWidth = this.chartContainer
         .getBoundingClientRect().width
-      const shouldResize = containerWidth !== currentContainerWidth
+      
+      const shouldResize = (containerWidth !== currentContainerWidth)
 
       if (shouldResize) {
         this.setState({
@@ -36,7 +36,7 @@ export default ChartComponent => (
     }
 
     renderChart() {
-      const parentWidth = this.state.containerWidth
+      const { containerWidth: parentWidth } = this.state
 
       return (
         <ChartComponent {...this.props} parentWidth={parentWidth} />
