+++ src/css/Theme.css
@@ -1,0 +1,7 @@
+/* .dark .theme { */
+    /* main color */
+    /* sub color */
+    /* text color */
+    /* shadow color? */
+    /*  */
+/* } */
